Added initial support for importing mzTab files
Former-commit-id: a40443ac66eadec551b412e82d4e6a93e3566a52

+++ src/main/java/net/sf/mzmine/project/impl/RawDataFileImpl.java
@@ -514,8 +514,10 @@ public class RawDataFileImpl implements RawDataFile, RawDataFileWriter {
 
     public synchronized void close() {
 	try {
-	    dataPointsFile.close();
-	    dataPointsFileName.delete();
+	    if(dataPointsFileName != null) {
+    	    	dataPointsFile.close();
+    	    	dataPointsFileName.delete();
+	    }
 	} catch (IOException e) {
 	    logger.warning("Could not close file " + dataPointsFileName + ": "
 		    + e.toString());
